LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event.https-expected.txt


PASS global setup
PASS Service Worker headers in the request of a fetch event
PASS Service Worker responds to fetch event with string
PASS Service Worker responds to fetch event using request fragment with string
PASS Service Worker responds to fetch event with blob body
PASS Service Worker responds to fetch event with the referrer URL
PASS Service Worker responds to fetch event with an existing client id
PASS Service Worker responds to fetch event with the correct resulting client id
PASS Service Worker does not respond to fetch event
PASS Service Worker responds to fetch event with null response body
PASS Service Worker fetches other file in fetch event
PASS Service Worker responds to fetch event with POST form
PASS Service Worker falls back to network in fetch event with POST form
PASS Multiple calls of respondWith must throw InvalidStateErrors
PASS Service Worker event.respondWith must set the used flag
PASS Service Worker should expose FetchEvent URL fragments.
FAIL Service Worker responds to fetch event with the correct cache types assert_equals: expected "no-cache" but got "default"
PASS Service Worker should intercept EventSource
PASS Service Worker responds to fetch event with the correct integrity_metadata
PASS FetchEvent#body is a string
FAIL FetchEvent#body is a ReadableStream promise_test: Unhandled rejection with value: object "NotSupportedError: ReadableStream uploading is not supported"
PASS FetchEvent#body is a string and is passed to network fallback
FAIL FetchEvent#body is a ReadableStream and is passed to network fallback promise_test: Unhandled rejection with value: object "NotSupportedError: ReadableStream uploading is not supported"
FAIL FetchEvent#body is a none Uint8Array ReadableStream and is passed to a service worker promise_rejects_js: function "function() { throw e }" threw object "NotSupportedError: ReadableStream uploading is not supported" ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS FetchEvent#body is a string, used and passed to network fallback
FAIL FetchEvent#body is a ReadableStream, used and passed to network fallback promise_rejects_js: function "function() { throw e }" threw object "NotSupportedError: ReadableStream uploading is not supported" ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS FetchEvent#body is a string, cloned and passed to network fallback
PASS FetchEvent#body is a ReadableStream, cloned and passed to network fallback
PASS FetchEvent#body is a blob
PASS FetchEvent#body is a blob and is passed to network fallback
PASS Service Worker responds to fetch event with the correct keepalive value
FAIL FetchEvent#request.isReloadNavigation is true (location.reload()) assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
FAIL FetchEvent#request.isReloadNavigation is true (history.go(0)) assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
FAIL FetchEvent#request.isReloadNavigation is true (POST + location.reload()) assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
FAIL FetchEvent#request.isReloadNavigation is true (with history traversal) assert_equals: expected "method = GET, isReloadNavigation = false" but got "method = GET, isReloadNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is true (with history.go(-1)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is true (with history.go(1)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is false (with history.go(0)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is false (with location.reload) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is true (with history.go(-2)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is true (with history.go(2)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
FAIL FetchEvent#request.isHistoryNavigation is true (POST + history.go(-1)) assert_equals: expected "method = GET, isHistoryNavigation = false" but got "method = GET, isHistoryNavigation = undefined"
PASS XHR upload progress events for response coming from SW
PASS XHR upload progress events for network fallback
PASS Fetch with POST with text on sw 421 response should not be retried.
PASS restore global state
